+++ data-raw/config_data_vars.R
@@ -1,5 +1,0 @@
-config_data_vars <- jsonlite::read_json(
-  "data-raw/config_data_vars.json",
-  simplifyVector = TRUE
-)
-usethis::use_data(config_data_vars, internal = TRUE, overwrite = TRUE)