+++ NEWS.md
@@ -1,3 +1,10 @@
+# preproc.iquizoo 1.3.0
+
+## Breaking Changes
+
+* Now internal data moved to {data.iquizoo} package.
+* Now `preproc()` accepts custom-defined functions and custom-defined configurations (input and output variable names) in a custom `.json` file.
+
 # preproc.iquizoo 1.2.1
 
 ## Breaking Changes